🔨 When install is enabled, don't generate static plugin binding
From cmake doc: https://cmake.org/cmake/help/v3.1/prop_tgt/INTERFACE_SOURCES.html
Targets with INTERFACE_SOURCES may not be exported with the export() or install(EXPORT) commands. This limitation may be lifted in a future version of CMake.

So it will be the user of Qaterial responsability to bind with correct plugins